Core: remove warnings and internalize some static variables - Remove unnecessary cast - Use getters/setters instead of public variables - Remove statement unnecessarily nested within else clause. - Remove duplicate checkings - Remove unused import - Remove unnecessary semicolon

+++ core/src/com/gsma/rcs/core/ims/protocol/rtp/core/RtcpSession.java
@@ -215,7 +215,7 @@ public class RtcpSession {
         }
 
         // get interval
-        t = (double) avgrtcpsize * n / bandwidth;
+        t = avgrtcpsize * n / bandwidth;
         if (t < rtcp_min_time)
             t = rtcp_min_time;
 
@@ -232,7 +232,7 @@ public class RtcpSession {
      * @param size
      */
     public void updateavgrtcpsize(int size) {
-        avgrtcpsize = (int) (0.0625 * (double) size + 0.9375 * (double) avgrtcpsize);
+        avgrtcpsize = (int) (0.0625 * size + 0.9375 * avgrtcpsize);
     }
 
     /**
